pyproject.toml
[tool.poetry]
name = "matcha-dl"
version = "0.1.0"
description = "Matcha-DL is a tool for supervised ontology alignment. It takes as input two ontologies and a set of mappings between them, and learns a model that can predict mappings between the two ontologies."
authors = ["Pedro Cotovio <pgcotovio@gmail.com>"]
readme = "README.md"
packages = [
    { include = "matcha_dl" },
]

include = ["matcha_dl/**/*"]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.3.0"
deeponto = "^0.9.1"
tensorboard = "^2.16.2"
dill = "^0.3.8"
lxml = "^5.2.1"
pandas = "^2.2.2"
pyyaml = "^6.0.1"
sentence-transformers = "^2.7.0"

[tool.poetry.scripts]
matchadl = "matcha_dl.delivery.cli:main"


[tool.poetry.group.prebuild.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.9.0"
pytest = "^8.1.1"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# isort
[tool.isort]
profile = "black"
# black
[tool.black]
line_length = 100
extend_skip = [".md", ".json"]
multi_line_output=3
force_grid_wrap = 0

# mypy
[tool.mypy]
warn_return_any = true
warn_unused_configs = true
files = ["tests/**/*.py"]